backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
opencv-python-headless
python-multipart
torch
torchvision
facenet-pytorch
pillow
numpy

